Fix v2 page titles
