+++ src/locales/en/messages.po
@@ -794,9 +794,9 @@ msgstr "Create an Ethereum address that can be used to pay your project directly
 msgid "Create payable address"
 msgstr "Create payable address"
 
-#: src/Router.tsx
 #: src/components/Projects/MyProjects.tsx
 #: src/components/Projects/index.tsx
+#: src/hooks/PageTitle.ts
 msgid "Create project"
 msgstr "Create project"
 
@@ -2073,8 +2073,8 @@ msgstr "Project {projectId}"
 msgid "Project {projectId} not found"
 msgstr "Project {projectId} not found"
 
-#: src/Router.tsx
 #: src/components/Navbar/MenuItems.tsx
+#: src/hooks/PageTitle.ts
 msgid "Projects"
 msgstr ""
