+++ src/locales/es/messages.po
@@ -799,9 +799,9 @@ msgstr "Crea una dirección Ethereum que pueda ser usada para pagar tu proyecto 
 msgid "Create payable address"
 msgstr "Crear dirección pagable"
 
-#: src/Router.tsx
 #: src/components/Projects/MyProjects.tsx
 #: src/components/Projects/index.tsx
+#: src/hooks/PageTitle.ts
 msgid "Create project"
 msgstr "Crear proyecto"
 
@@ -2078,8 +2078,8 @@ msgstr "Proyecto {projectId}"
 msgid "Project {projectId} not found"
 msgstr "Proyecto {projectId} no encontrado"
 
-#: src/Router.tsx
 #: src/components/Navbar/MenuItems.tsx
+#: src/hooks/PageTitle.ts
 msgid "Projects"
 msgstr "Proyectos"
