+++ src/locales/fr/messages.po
@@ -799,9 +799,9 @@ msgstr "Créez une adresse Ethereum qui peut être utilisée pour payer directem
 msgid "Create payable address"
 msgstr "Créer une adresse de paiement"
 
-#: src/Router.tsx
 #: src/components/Projects/MyProjects.tsx
 #: src/components/Projects/index.tsx
+#: src/hooks/PageTitle.ts
 msgid "Create project"
 msgstr "Créer un projet"
 
@@ -2078,8 +2078,8 @@ msgstr "Projet {projectId}"
 msgid "Project {projectId} not found"
 msgstr "Projet {projectId} introuvable"
 
-#: src/Router.tsx
 #: src/components/Navbar/MenuItems.tsx
+#: src/hooks/PageTitle.ts
 msgid "Projects"
 msgstr "Projets"
