+++ src/locales/ru/messages.po
@@ -799,9 +799,9 @@ msgstr "Создать Ethereum-адрес, который можно испол
 msgid "Create payable address"
 msgstr "Создать адреса для оплаты"
 
-#: src/Router.tsx
 #: src/components/Projects/MyProjects.tsx
 #: src/components/Projects/index.tsx
+#: src/hooks/PageTitle.ts
 msgid "Create project"
 msgstr "Создать проект"
 
@@ -2078,8 +2078,8 @@ msgstr ""
 msgid "Project {projectId} not found"
 msgstr "Проект {projectId} не найден"
 
-#: src/Router.tsx
 #: src/components/Navbar/MenuItems.tsx
+#: src/hooks/PageTitle.ts
 msgid "Projects"
 msgstr "Проекты"
