+++ src/locales/tr/messages.po
@@ -799,9 +799,9 @@ msgstr "Projenize doğrudan ödeme yapmak için kullanılabilecek bir Ethereum a
 msgid "Create payable address"
 msgstr "Ödeme adresi oluşturun"
 
-#: src/Router.tsx
 #: src/components/Projects/MyProjects.tsx
 #: src/components/Projects/index.tsx
+#: src/hooks/PageTitle.ts
 msgid "Create project"
 msgstr "Proje oluştur"
 
@@ -2078,8 +2078,8 @@ msgstr "{projectId} projesi"
 msgid "Project {projectId} not found"
 msgstr "Proje {projectId} bulunamadı"
 
-#: src/Router.tsx
 #: src/components/Navbar/MenuItems.tsx
+#: src/hooks/PageTitle.ts
 msgid "Projects"
 msgstr "Projeler"
