+++ src/locales/zh/messages.po
@@ -799,9 +799,9 @@ msgstr "创建一个可用于直接支付给你的项目的以太坊地址。"
 msgid "Create payable address"
 msgstr "创建可支付地址"
 
-#: src/Router.tsx
 #: src/components/Projects/MyProjects.tsx
 #: src/components/Projects/index.tsx
+#: src/hooks/PageTitle.ts
 msgid "Create project"
 msgstr "创建项目"
 
@@ -2078,8 +2078,8 @@ msgstr "项目 {projectId}"
 msgid "Project {projectId} not found"
 msgstr "项目 {projectId} 未找到"
 
-#: src/Router.tsx
 #: src/components/Navbar/MenuItems.tsx
+#: src/hooks/PageTitle.ts
 msgid "Projects"
 msgstr "项目"
